Issue: #43 Link: https://github.com/loopj/jquery-simple-slider/issues/43
Users were confused by documentation showing data-slider-* attributes as being
the way to configure the slider.  data-slider-* options were being processed by
a visitor function that for each input created a settings object and invoked
SimpleSlider with those settings.

Moved the logic that created a settings object to
SimpleSlider.loadDataSettings(). If SimpleSlider() is called this method will
be invoked to create the settings from the data-slider-* attributes. If
SimpleSlider(settings) is called it will not (settings will come from the
object being passed).